rename: 폴더명 케밥케이스로 수정
